backend/requirements.txt
# Core Framework
Flask==3.0.0
Flask-CORS==4.0.0

# Supabase (will auto-install compatible postgrest)
supabase==2.9.1

# HTTP Client
httpx==0.27.2

# JWT Handling
PyJWT==2.8.0

# Environment Variables
python-dotenv==1.0.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3

# Development
black==23.12.1
flake8==7.0.0

# WebSockets
websockets==13.1